Move signup completion controller logic
회원가입 완료 로직을 user/role/role.controller.ts 로 이동
